requirements.txt
#venv/Scripts/activate
fastapi[all] 
tortoise-orm 
pydantic 
pytest
aiofiles